+++ packages/playwright-toolkit/Classes/Database/Driver/TestDatabaseDriver.php
@@ -16,6 +16,11 @@ interface TestDatabaseDriver
     /**
      * @return array<string, mixed>
      */
+    public function baseConnectionOverrides(): array;
+
+    /**
+     * @return array<string, mixed>
+     */
     public function templateConnectionOverrides(): array;
 
     /**